# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_kishert-lab_taxi-platform_internal_domain.PaymentMethod:
    enum:
    - cash
    - card
    - corporate
    type: string
    x-enum-varnames:
    - PaymentMethodCash
    - PaymentMethodCard
    - PaymentMethodCorporate
...
  github_com_kishert-lab_taxi-platform_internal_dto.MoneyResponse:
    properties:
      amount:
        example: 25000
        type: integer
      currency:
        example: RUB
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.OrderPricingResponse:
    properties:
      assigned_tariff_id:
        example: 22222222-2222-2222-2222-222222222222
        type: string
      assigned_taxi_park_id:
        example: 55555555-5555-5555-5555-555555555555
        type: string
      estimated_price:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.MoneyResponse'
      estimated_price_max:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.MoneyResponse'
      estimated_price_min:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.MoneyResponse'
      estimated_price_source:
        example: car_class_catalog
        type: string
      final_price:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.MoneyResponse'
      is_final:
        type: boolean
      message:
        example: Цена будет рассчитана после назначения водителя
        type: string
      price_available:
        type: boolean
      pricing_mode:
        example: distance_time
        type: string
      route_data_version:
        type: string
      route_distance_meters:
        example: 4200
        type: integer
      route_duration_seconds:
        example: 660
        type: integer
      route_source:
        example: osrm
        type: string
      search_radius_meters:
        example: 5000
        type: integer
      taxi_park_count:
        example: 3
        type: integer
      unavailable_reason:
        example: routing_unavailable
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerMePatchRequest:
    properties:
      avatar_url:
        example: https://cdn.example.com/passengers/111/avatar.jpg
        type: string
      email:
        example: test@example.com
        type: string
      name:
        example: Сергей
        maxLength: 255
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerMeResponse:
    properties:
      avatar_url:
        example: https://cdn.example.com/passengers/111/avatar.jpg
        type: string
      email:
        example: test@example.com
        type: string
      id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
      name:
        example: Сергей
        type: string
      phone:
        example: "+79124966126"
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenRequest:
    properties:
      device_id:
        example: pixel-8-pro
        type: string
      platform:
        enum:
        - android
        - ios
        - web
        example: android
        type: string
      token:
        example: fcm_device_token
        type: string
    required:
    - platform
    - token
    type: object
  github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenResponse:
    properties:
      device_id:
        example: pixel-8-pro
        type: string
      platform:
        example: android
        type: string
      token:
        example: fcm_device_token
        type: string
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.TaxiParkCreateOrderRequest:
    properties:
      comment:
        example: Entrance 2
        type: string
      destination_address:
        example: Lenina 50
        type: string
      destination_location:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.TaxiParkOrderCoordinatesRequest'
      passenger_name:
        example: Irina
        type: string
      passenger_phone:
        example: "+79990000000"
        type: string
      payment_method:
        allOf:
        - $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_domain.PaymentMethod'
        example: cash
      payment_type:
        allOf:
        - $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_domain.PaymentMethod'
        example: cash
      pickup_address:
        example: Mira 8
        type: string
      pickup_location:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.TaxiParkOrderCoordinatesRequest'
      tariff_id:
        example: 4e99f0be-5ea1-4353-8307-d1555f588825
        type: string
    required:
    - tariff_id
    type: object
...
  github_com_kishert-lab_taxi-platform_internal_dto.TaxiParkOrderCoordinatesRequest:
    properties:
      latitude:
        example: 56.835128
        type: number
      longitude:
        example: 60.598698
        type: number
    type: object
...
  github_com_kishert-lab_taxi-platform_pkg_response.Error:
    properties:
      error:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
  github_com_kishert-lab_taxi-platform_pkg_response.ErrorBody:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.ErrorCode'
        example: ORDER_NOT_FOUND
      details:
        additionalProperties: {}
        type: object
      message:
        example: Order not found
        type: string
    type: object
  github_com_kishert-lab_taxi-platform_pkg_response.ErrorCode:
    enum:
    - VALIDATION_ERROR
    - INVALID_PHONE
    - INVALID_CODE
    - CODE_EXPIRED
    - TOO_MANY_ATTEMPTS
    - PASSENGER_BLOCKED
    - TOKEN_EXPIRED
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - ORDER_NOT_FOUND
    - ORDER_INVALID_STATE
    - DRIVER_NOT_AVAILABLE
    - ORDER_ALREADY_ASSIGNED
    - DISPATCH_IN_PROGRESS
    - RATE_LIMITED
    - CONSENT_REQUIRED
    - NOT_IMPLEMENTED
    - INTERNAL_ERROR
    type: string
    x-enum-varnames:
    - CodeValidationError
    - CodeInvalidPhone
    - CodeInvalidCode
    - CodeCodeExpired
    - CodeTooManyAttempts
    - CodePassengerBlocked
    - CodeTokenExpired
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeOrderNotFound
    - CodeOrderInvalidState
    - CodeDriverNotAvailable
    - CodeOrderAlreadyAssigned
    - CodeDispatchInProgress
    - CodeRateLimited
    - CodeConsentRequired
    - CodeNotImplemented
    - CodeInternalError
  github_com_kishert-lab_taxi-platform_pkg_response.Meta:
    properties:
      request_id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
    type: object
...
  internal_transport_http_handler.PassengerMeSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerMeResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.PassengerPushTokenSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
  internal_transport_http_handler.TaxiParkEstimateResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.OrderPricingResponse'
      meta:
        $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Meta'
    type: object
...
info:
  contact:
    email: support@example.com
    name: Taxi Platform API Support
  description: Production-grade taxi backend API for small cities and rural areas.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Taxi Platform API
  version: 0.1.0
paths:
...
  /passenger/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerMeSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Get current passenger profile
      tags:
      - passenger
    patch:
      consumes:
      - application/json
      parameters:
      - description: Passenger profile patch
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerMePatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerMeSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Update current passenger profile
      tags:
      - passenger
...
  /passenger/push-tokens:
    post:
      consumes:
      - application/json
      description: Registers or updates the push token for the authenticated passenger
        device.
      parameters:
      - description: Passenger push token registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerPushTokenSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Register passenger push token
      tags:
      - passenger
  /passenger/push/token:
    post:
      consumes:
      - application/json
      description: Registers or updates the push token for the authenticated passenger
        device.
      parameters:
      - description: Passenger push token registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.PassengerPushTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.PassengerPushTokenSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Register passenger push token
      tags:
      - passenger
...
  /taxi-park/orders/estimate:
    post:
      consumes:
      - application/json
      description: Uses the selected authorized park tariff, preserving fixed/distance/time/combined
        modes. Snapshot amounts are kopecks. OSRM duration excludes live traffic.
        Does not create an order.
      parameters:
      - description: Order form
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kishert-lab_taxi-platform_internal_dto.TaxiParkCreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_handler.TaxiParkEstimateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_kishert-lab_taxi-platform_pkg_response.Error'
      security:
      - BearerAuth: []
      summary: Estimate a dispatcher order using road distance and duration
      tags:
      - taxi-park-orders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer ` prefix, for example: Bearer eyJhbGciOi...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"